% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustrfiltr.R
\name{clustrfiltr}
\alias{clustrfiltr}
\title{Filter clusters based on their gene-set size}
\usage{
clustrfiltr(getclustrs_data, size_filtr = 3)
}
\arguments{
\item{getclustrs_data}{A \code{dataframe} of type \emph{t} that typically corresponds to the output of the \code{getclustrs()} function. This input holds at least the columns named \code{gene_id} and \code{clustr} respectively holding Ensembl gene and cluster identifiers for the deregulated genes.}

\item{size_filtr}{The minimum number of genes required for a cluster to be retained (by default: 3).}
}
\value{
A named \code{list} holding 2 components, where :
- \code{kept} is a dataframe of type \emph{t} similar to the \emph{getclustrs_data} dataframe input with the rows kept after the filter (the genes are part of a cluster equal and over the size limit)
-\code{removed} is a dataframe of type \emph{t} similar to the \emph{getclustrs_data} dataframe input with the rows removed after the filter (the genes are part of clusters under the size limit)
- \code{params} is a list of the main parameters used
}
\description{
This function filters clusters based on size, selecting genes that belong to clusters meeting the user-defined size criteria.
}
