% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curves_to_pdf.R
\name{curves_to_pdf}
\alias{curves_to_pdf}
\title{Generate curvesplots per-cluster to PDF}
\usage{
curves_to_pdf(
  lonelyfishing_data,
  bmdboot_data,
  clustrfusion_data,
  id_col_for_curves = "id",
  tested_doses,
  annot_order,
  ...,
  xunit,
  xtitle,
  ytitle,
  colors = c(inc = "#1B9E77", dec = "#D95F02", U = "#7570B3", bell = "#E7298A"),
  path,
  output_filename,
  overwrite = TRUE
)
}
\arguments{
\item{lonelyfishing_data}{The named \code{list} output of the \code{lonelyfishing()} function.}

\item{bmdboot_data}{The DRomics bmdboot dataframe results after DRomics::bmdfilter()}

\item{clustrfusion_data}{The named \code{list} output of the \code{clustrfusion()} function.}

\item{id_col_for_curves}{The column giving the identification of each curve (default is "id")}

\item{tested_doses}{A vector of the tested doses that can be found in the output of the \code{DRomics::drcfit()}function (unique(f$omicdata$dose))}

\item{annot_order}{A vector specifying the prioritized order of annotation sources used in the \code{clustrenrich()} function. This order determines which term or pathway will be used as primary descriptor for each cluster. The relevance of each source can be based on the quantity or quality of information they provide, tailored to the specific case study. For example, for Danio rerio, you might prioritize sources like c("GO:BP", "KEGG", "WP") based on the abundance of information they offer.}

\item{...}{Additional arguments passed to the \code{DRomics::curvesplot()} function}

\item{xunit}{Unit for the x scale}

\item{xtitle}{X-axis title}

\item{ytitle}{Y-axis title}

\item{colors}{A vector of colors for different trends (default is set at c("inc" = "#1B9E77", "dec" = "#D95F02", "U" = "#7570B3", "bell" = "#E7298A"))}

\item{path}{Destination folder for the output data results.}

\item{output_filename}{Output PDF filename.}

\item{overwrite}{If \code{TRUE}, the function overwrites existing output files; otherwise, it reads the existing file. (default is set to \code{FALSE}).}
}
\value{
No return value; the \code{.pdf} file is downloaded and stored in the specified director. It holds a curvesplot per slide for each cluster, and those for the Friendly (if created), Lonely and All deregulated gene transcripts.
}
\description{
This function generates a PDF file featuring dose-response curves for each cluster of deregulated genes. Each plot is labeled with the cluster ID and the number of transcripts in the cluster. By default, the curves are color-coded to indicate whether the trend is increasing, decreasing, U-shaped, or bell-shaped. Users can also apply any modifications available in the original DRomics::curvesplot() function.
}
