% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getclustrs.R
\name{getclustrs}
\alias{getclustrs}
\title{Retreive and integrate the clustered Protein-Protein Interaction Network (PPIN) data}
\usage{
getclustrs(gene_data, colname_for_merge, path, nodetable_filename)
}
\arguments{
\item{gene_data}{A dataframe of type \emph{t} that typically corresponds to the output of the \code{getids()} or \code{getregs()} function. This input holds at least one column named "gene_id” holding gene identifiers for the deregulated genes.}

\item{colname_for_merge}{The identifier column used as query in order to create the PPIN in Cytoscape using the StringApp (e.g. "ensembl_gene_id" or "uniprotsptrembl")}

\item{path}{Folder where the nodetable exported from Cytoscape is found}

\item{nodetable_filename}{Filename of the nodetable}
}
\value{
A \code{dataframe} of type \emph{t} similar to the \emph{gene_data} dataframe input with an added column indicating to which cluster belongs each gene (if no cluster is associated : NA)
}
\description{
This function retrieves and reformats the clustered PPIN data from the StringApp in Cytoscape.
}
