% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getregs.R
\name{getregs}
\alias{getregs}
\title{Retrieve the regulatory status of deregulated genes}
\usage{
getregs(getids_data, regulator_file, coregulator_file)
}
\arguments{
\item{getids_data}{A dataframe of type \emph{t} that typically corresponds to the output of the \code{getids()} function. This input holds at least one column named "gene_id" holding gene identifiers for the deregulated genes.  It should be a subset of the output of the \code{getids()} function with only the significantly deregulated transcripts as rows and transcript/gene identifiers as columns.}

\item{regulator_file}{The previously downloaded \code{.txt} Transcription Factor file
from the AnimalTFDB4 database (https://guolab.wchscu.cn/AnimalTFDB4/#/).}

\item{coregulator_file}{The previously downloaded \code{.txt} Transcription Co-Factor file from the AnimalTFDB4 database (https://guolab.wchscu.cn/AnimalTFDB4/#/).}
}
\value{
A \code{dataframe} of type \emph{t} similar to the \emph{getids_data} input with an added column indicating the (co-)regulator status of each transcript's deregulated gene.
}
\description{
This function retrieves information on the regulatory status of genes using the files downloaded with the \code{dl_regulation_data()} function. It firstly checks if the specifed files have beel downloaded, and whether they contain the è \code{Ensembl}
}
\examples{
\dontrun{
# Create example gene data (normally from getids())
example_getids_res <- data.frame(
  transcript_id = c("ENSDART00000000069.8", "ENSDART00000002164.9", 
                    "ENSDART00000001691.8", "ENSDART00000000070.7"),
  gene_id = c("ENSDARG00000000068", "ENSDARG00000000069",
              "ENSDARG00000001463", "ENSDARG00000008433")
  gene_name = c("NHERF1", "DAP", "tdh2", "UNC45B"),
  stringsAsFactors = FALSE
)

# Add regulatory status (requires downloaded TF/CoTF files)
gene_regs <- getregs(
  getids_data = example_getids_res,
  regulator_file = file.path(tempdir(), "Danio_rerio_TF_example.txt"),
  coregulator_file = file.path(tempdir(), "Danio_rerio_Cof_example.txt")
)
}
}
