% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_qmd.R
\name{render_qmd}
\alias{render_qmd}
\title{Render and Move Quarto Documents}
\usage{
render_qmd(input_file, output_file, output_path, file_ext, ...)
}
\arguments{
\item{input_file}{A character string specifying the path to the input Quarto \code{.qmd} file.}

\item{output_path}{A character string specifying the directory where the rendered file should be moved.}

\item{...}{Additional arguments passed to \code{quarto::quarto_render}.}
}
\value{
This function does not return a value. It performs rendering and file moving operations, and displays a message indicating the file movement.
}
\description{
This function renders a Quarto \code{.qmd} document to a specified output format and then moves
the resulting file to a designated output path. It uses \code{quarto::quarto_render} to
perform the rendering and \code{fs::file_move} to handle the file movement.
}
