% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_to_csv.R
\name{results_to_csv}
\alias{results_to_csv}
\title{Generate a workflow summary CSV file}
\usage{
results_to_csv(
  lonelyfishing_data,
  bmdboot_data,
  path,
  output_filename,
  overwrite = TRUE
)
}
\arguments{
\item{lonelyfishing_data}{The named \code{list} output of the \code{lonelyfishing()} function.}

\item{bmdboot_data}{The DRomics bmdboot \code{dataframe} results after DRomics::bmdfilter()}

\item{path}{Destination folder for the output data results.}

\item{output_filename}{Output CSV filename.}

\item{overwrite}{If \code{TRUE}, the function overwrites existing output files; otherwise, it reads the existing file. (default is set to \code{FALSE}).}
}
\value{
No return value; the \code{.csv} file is downloaded and stored in the specified director. It holds the results of the final step of the workflow with each row being a combination of transcript, gene and other identifiers, cluster and biological function annotation. This is to be used as a support for exploration and mechanism discovery.
}
\description{
This function creates a concise summary table capturing key details derived from the workflow results from the \code{lonelyfishing()} function. It is formatted with each row being a combination of transcript, cluster and biological function annotation.
}
